deploy/requirements.txt
discord.py==2.5.2
aiosqlite==0.21.0
aiohttp>=3.7.4
Pillow>=8.0.0
python-dotenv>=0.19.0
requests>=2.25.0
beautifulsoup4>=4.9.0
pytz>=2021.1
